global cleanup will delete iba dat files only RMI #10528
svn path=/trunk/; revision=456

+++ ibaDatCoordinator/Processing/FileQuotaCleanup.cs
@@ -10,14 +10,14 @@ using System.Threading;
 
 namespace iba.Processing
 {
-    public class FileQuotaCleanup
+    internal class FileQuotaCleanup
     {
-        private CleanupTaskData m_task;
         private LinkedList<String> m_files;
-        private bool m_fastSearch;
-        string m_extension;
-        UInt64 m_size;
-        CancellationToken m_cancelToken = new CancellationToken(false);
+        protected CleanupTaskData m_task;
+        protected string m_extension;
+        protected UInt64 m_size;
+        protected CancellationToken m_cancelToken = new CancellationToken(false);
+        protected bool m_failureWhilePreviouslyScanning = false;
 
         public FileQuotaCleanup(CleanupTaskData task, string extension)
         {
@@ -30,12 +30,6 @@ namespace iba.Processing
             : this(task, extension)
         {
             m_cancelToken = ct;
-        }
-
-        public bool FastSearch
-        {
-            get { return m_fastSearch; }
-            set { m_fastSearch = value; }
         }
 
         public void AddFile(string filename)
@@ -80,7 +74,7 @@ namespace iba.Processing
             m_task = newTask;
         }
 
-        struct DateAndName
+        protected struct DateAndName
         {
             public DateTime time;
             public string filename;
@@ -92,11 +86,10 @@ namespace iba.Processing
             }
         };
 
-        private bool FailureWhilePreviouslyScanning;
 
         private void Reset()
         {
-            FailureWhilePreviouslyScanning = false;
+            m_failureWhilePreviouslyScanning = false;
             m_size = 0;
             m_files.Clear();
             Log(iba.Logging.Level.Info, iba.Properties.Resources.determiningQuota, "");
@@ -106,11 +99,7 @@ namespace iba.Processing
             var sw = Stopwatch.StartNew();
             try
             {
-                if (FastSearch)
-                    TraverseFilesystemFastSearch(DateAndNames);
-                else
-                    TraverseFileSystemClassic(DateAndNames);
-
+                TraverseFileSystem(DateAndNames);
 
                 if (m_cancelToken.IsCancellationRequested)
                     return;
@@ -129,7 +118,7 @@ namespace iba.Processing
             catch (Exception ex)
             {
                 Log(iba.Logging.Level.Exception, iba.Properties.Resources.logCleanupTallyingErrorGeneral + ex.Message, "");
-                FailureWhilePreviouslyScanning = true;
+                m_failureWhilePreviouslyScanning = true;
                 m_files.Clear();
             }
 
@@ -138,7 +127,7 @@ namespace iba.Processing
             Log(iba.Logging.Level.Debug, String.Format("{0} files with size of {1}", m_files.Count, PathUtil.GetSizeReadable((long)m_size)), "");
         }
 
-        private void TraverseFileSystemClassic(List<DateAndName> DateAndNames)
+        protected virtual void TraverseFileSystem(List<DateAndName> DateAndNames)
         {
             foreach (string dir in Directory.GetDirectories(m_task.DestinationMapUNC))
             {
@@ -160,14 +149,14 @@ namespace iba.Processing
                         catch (Exception ex)
                         {
                             Log(iba.Logging.Level.Exception, String.Format(iba.Properties.Resources.logCleanupTallyingErrorFile, dir) + ex.Message, "");
-                            FailureWhilePreviouslyScanning = true;
+                            m_failureWhilePreviouslyScanning = true;
                         }
                     }
                 }
                 catch (Exception ex)
                 {
                     Log(iba.Logging.Level.Exception, iba.Properties.Resources.logCleanupTallyingErrorDir + " " + ex.Message, "");
-                    FailureWhilePreviouslyScanning = true;
+                    m_failureWhilePreviouslyScanning = true;
                 }
             }
             foreach (string file in Utility.PathUtil.GetFilesMultipleExtensions(m_task.DestinationMapUNC, "*" + m_extension, SearchOption.TopDirectoryOnly))
@@ -185,35 +174,13 @@ namespace iba.Processing
                 catch (Exception ex)
                 {
                     Log(iba.Logging.Level.Exception, iba.Properties.Resources.logCleanupTallyingErrorGeneral + " " + ex.Message, "");
-                    FailureWhilePreviouslyScanning = true;
-                }
-            }
-        }
-
-        private void TraverseFilesystemFastSearch(List<DateAndName> DateAndNames)
-        {
-            var dir = String.Empty;
-            foreach (var inf in FastDirectoryEnumerator.EnumerateFiles(m_task.DestinationMapUNC, "*" + m_extension, SearchOption.AllDirectories))
-            {
-                try
-                {
-                    if (m_cancelToken.IsCancellationRequested)
-                        return;
-
-                    m_size += (ulong)inf.Length;
-                    DateAndNames.Add(new DateAndName(inf.FullName, inf.LastWriteTime));
-
-                }
-                catch (Exception ex)
-                {
-                    Log(iba.Logging.Level.Exception, String.Format(iba.Properties.Resources.logCleanupTallyingErrorFile, dir) + ex.Message, "");
-                    FailureWhilePreviouslyScanning = true;
+                    m_failureWhilePreviouslyScanning = true;
                 }
             }
         }
 
 
-        private void Log(Logging.Level level, string message, string datfile)
+        protected void Log(Logging.Level level, string message, string datfile)
         {
             LogExtraData data = new LogExtraData(datfile, m_task, m_task.ParentConfigurationData);
             if (LogData.Data.Logger != null && LogData.Data.Logger.IsOpen)
@@ -253,19 +220,19 @@ namespace iba.Processing
 
         public void Clean(string datfile) //the datfile parameter here is used for logging, nothing else
         {
-            if (FailureWhilePreviouslyScanning)
+            if (m_failureWhilePreviouslyScanning)
             {
                 Reset();
-                if (FailureWhilePreviouslyScanning) return; //fix next time.
+                if (m_failureWhilePreviouslyScanning) return; //fix next time.
             }
 
             string origPath = Path.GetFullPath(m_task.DestinationMapUNC);
 
             var quota = GetQuota();
 
-            var bytesToDelete = 0L;
-            if (m_size > quota)
-                bytesToDelete = (long)(m_size - quota);
+            //var bytesToDelete = 0L;
+            //if (m_size > quota)
+            //    bytesToDelete = (long)(m_size - quota);
 
             //try
             //{
@@ -323,29 +290,39 @@ namespace iba.Processing
                             return; //forget about cleanup until problem is fixed.
                         }
                     }
+
+
                     FileInfo inf = new FileInfo(file);
                     DirectoryInfo parent = inf.Directory;
                     ulong size = (ulong)inf.Length;
+
                     try
                     {
+                        if (IsIbaDatfile(file))
+                        {
+                            File.Delete(file);
 
-                        File.Delete(file);
+                            if (size > m_size)
+                                m_size = 0;
+                            else
+                                m_size -= size;
 
-                        if (size > m_size)
-                            m_size = 0;
+                            string message = String.Format(iba.Properties.Resources.logCleanupSuccessRemoveFile, file);
+                            // string message = String.Format("Deleted File Time: {2} File: {0} Size: {1} ", file, PathUtil.GetSizeReadable(inf.Length), inf.LastWriteTime);
+                            Log(iba.Logging.Level.Info, message, datfile);
+                            //Log(iba.Logging.Level.Exception, String.Format("size before deleting {0}, filesize {1}, difference {2}", m_size, size, m_size - size), file);
+                        }
                         else
-                            m_size -= size;
-
-                        string message = String.Format(iba.Properties.Resources.logCleanupSuccessRemoveFile, file);
-                        // string message = String.Format("Deleted File Time: {2} File: {0} Size: {1} ", file, PathUtil.GetSizeReadable(inf.Length), inf.LastWriteTime);
-                        Log(iba.Logging.Level.Info, message, datfile);
-                        //Log(iba.Logging.Level.Exception, String.Format("size before deleting {0}, filesize {1}, difference {2}", m_size, size, m_size - size), file);
+                        {
+                            string message = String.Format(iba.Properties.Resources.logCleanupSkippedNonIbaDatFile, file);
+                            Log(iba.Logging.Level.Info, message, datfile);
+                        }
                     }
                     catch (Exception ex)
                     {
                         string message = String.Format(iba.Properties.Resources.logCleanupErrorRemoveFile, file, ex.Message);
                         Log(iba.Logging.Level.Warning, message, datfile);
-                        throw ex;
+                        throw;
                     }
 
                     while (DirectoryEmpty(parent.FullName) && parent.FullName != origPath)
@@ -361,7 +338,7 @@ namespace iba.Processing
                         {
                             string message = String.Format(iba.Properties.Resources.logCleanupErrorRemoveDirectory, parent.FullName, ex.Message);
                             Log(iba.Logging.Level.Warning, message, datfile);
-                            throw ex;
+                            throw;
                         }
                         parent = grandParent;
                     }
@@ -382,6 +359,12 @@ namespace iba.Processing
             //    string message = String.Format("Quota Not Exceeded: filesize: {0}  Quota: {1}", m_size, m_task.Quota);
             //    Log(iba.Logging.Level.Warning, message, datfile);
             //}
+        }
+
+
+        protected virtual bool IsIbaDatfile(string file)
+        {
+            return true;
         }
 
         [StructLayout(LayoutKind.Sequential)]
